+++ Source/ITI/DDD/Domain/DbValueObject.cs
@@ -1,4 +1,5 @@
-﻿using Microsoft.EntityFrameworkCore;
+﻿using AutoMapper;
+using Microsoft.EntityFrameworkCore;
 
 namespace ITI.DDD.Domain.ValueObjects
 {