+++ Source/Tests/TestApp/DataContext/DataModel/DbFacility.cs
@@ -1,5 +1,6 @@
 ﻿using ITI.Baseline.Audit;
 using ITI.DDD.Infrastructure.DataContext;
+using System;
 using System.ComponentModel.DataAnnotations;
 using TestApp.Domain.ValueObjects;
 
@@ -7,6 +8,13 @@ namespace TestApp.DataContext.DataModel
 {
     public class DbFacility : DbEntity, IDbAudited
     {
+#nullable disable
+        [Obsolete("EF cannot map navigation properties in constructor", true)]
+        protected DbFacility()
+        {
+        }
+#nullable enable
+
         public DbFacility(string? name, FacilityContact contact)
         {
             Name = name;